Revert "refactor: Layout component update"
This reverts commit d1797620fbf4cc6f3124923720fb91f3c38add78.
