+++ pages/_app.tsx
@@ -1,6 +1,6 @@
 import '../styles/globals.css'
 import type { AppProps } from 'next/app'
-import Layout from '../components/views/Layout'
+import Layout from './api/Layout'
 
 function MyApp({ Component, pageProps }: AppProps) {
   return <Layout>
