+++ pages/index.tsx
@@ -7,6 +7,8 @@ const Home: NextPage = (props) => {
         main
         h-screen
         w-full
+        bg-light-grey
+        dark:bg-midnight 
       "
     >
     </div>
